daemon: thin tasks cli shell
