+++ crates/daemon/src/tasks_cli.rs
@@ -15,6 +15,17 @@ use loong_spec::CliResult;
 use serde_json::{Value, json};
 use std::path::PathBuf;
 
+#[path = "tasks_cli_render.rs"]
+mod render_support;
+#[path = "tasks_cli_status.rs"]
+mod status_support;
+
+use self::render_support::render_tasks_cli_text;
+use self::status_support::{
+    TaskStatusSummary, build_task_status_payload, summarize_task_status_payload,
+    unknown_task_status_payload,
+};
+
 #[derive(Subcommand, Debug, Clone, PartialEq, Eq)]
 pub enum TasksCommands {
     /// Queue one async background task on top of the current session runtime
@@ -1156,30 +1167,6 @@ fn ensure_background_task_status_payload(status_payload: &Value, task_id: &str) 
     Ok(())
 }
 
-fn summarize_task_status_payload(status_payload: &Value) -> CliResult<TaskStatusSummary> {
-    let session = status_payload
-        .get("session")
-        .ok_or_else(|| "task status payload missing session object".to_owned())?;
-    let delegate = status_payload
-        .get("delegate_lifecycle")
-        .cloned()
-        .unwrap_or(Value::Null);
-    let session_kind = session.get("kind").and_then(Value::as_str).unwrap_or("");
-    let delegate_mode = delegate.get("mode").and_then(Value::as_str).unwrap_or("");
-    let staleness_state = delegate
-        .get("staleness")
-        .and_then(|value| value.get("state"))
-        .and_then(Value::as_str)
-        .unwrap_or("");
-    let is_background_task = session_kind == "delegate_child" && delegate_mode == "async";
-    let is_overdue = staleness_state == "overdue";
-
-    Ok(TaskStatusSummary {
-        is_background_task,
-        is_overdue,
-    })
-}
-
 fn parse_task_state_filter(raw_state: &str) -> CliResult<mvp::session::repository::SessionState> {
     match raw_state {
         "ready" => Ok(mvp::session::repository::SessionState::Ready),
@@ -1191,347 +1178,11 @@ fn parse_task_state_filter(raw_state: &str) -> CliResult<mvp::session::repositor
     }
 }
 
-#[derive(Debug, Clone, Copy)]
-struct TaskStatusSummary {
-    is_background_task: bool,
-    is_overdue: bool,
-}
-
 #[derive(Debug, Clone, PartialEq, Eq)]
 struct ResolvedCliTaskTarget {
     task_id: String,
     owner_session_id: String,
     task_session_id: String,
-}
-
-fn build_task_status_payload(
-    session: &Value,
-    delegate: &Value,
-    task_progress: &Value,
-    terminal_outcome_state: &Value,
-    recovery: &Value,
-    approval_requests: &Value,
-    approval_attention_summary: &Value,
-    tool_policy: &Value,
-    recent_events: &Value,
-) -> Value {
-    let session_state = session
-        .get("state")
-        .and_then(Value::as_str)
-        .unwrap_or("unknown");
-    let phase = delegate
-        .get("phase")
-        .and_then(Value::as_str)
-        .unwrap_or("unknown");
-    let staleness_state = delegate
-        .get("staleness")
-        .and_then(|value| value.get("state"))
-        .and_then(Value::as_str);
-    let cancellation_state = delegate
-        .get("cancellation")
-        .and_then(|value| value.get("state"))
-        .and_then(Value::as_str);
-    let approval_attention_count = approval_attention_summary
-        .get("needs_attention_count")
-        .and_then(Value::as_u64)
-        .unwrap_or(0);
-    let has_approval_attention = approval_attention_count > 0;
-    let approval_primary_action = primary_approval_action(approval_requests).map(ToOwned::to_owned);
-    let recovered = recent_events_contains_kind(recent_events, "delegate_recovery_applied");
-    let tool_narrowing_active = task_tool_narrowing_active(tool_policy);
-    let task_progress_status = task_progress
-        .get("status")
-        .and_then(Value::as_str)
-        .filter(|value| !value.trim().is_empty());
-    let task_progress_handle_state = task_progress
-        .get("active_handles")
-        .and_then(Value::as_array)
-        .and_then(|handles| handles.first())
-        .and_then(|handle| handle.get("state"))
-        .and_then(Value::as_str)
-        .filter(|value| !value.trim().is_empty());
-    let terminal_outcome_state = terminal_outcome_state
-        .as_str()
-        .filter(|value| !value.trim().is_empty());
-    let recovery_kind = recovery
-        .get("kind")
-        .and_then(Value::as_str)
-        .filter(|value| !value.trim().is_empty());
-    let kind = derive_task_status_kind(
-        session_state,
-        phase,
-        staleness_state,
-        cancellation_state,
-        has_approval_attention,
-        task_progress_status,
-        task_progress_handle_state,
-        terminal_outcome_state,
-        recovery_kind,
-    );
-    let display = render_task_status_display(kind, recovered);
-    let blocked = task_status_is_blocked(kind);
-    let terminal = task_status_is_terminal(kind);
-    let status = kind;
-    let needs_attention = task_status_needs_attention(kind, approval_primary_action.as_deref());
-    let next_action = task_status_next_action(kind, approval_primary_action.as_deref());
-    let signals = build_task_status_signals(
-        kind,
-        recovered,
-        tool_narrowing_active,
-        has_approval_attention,
-        staleness_state,
-        cancellation_state,
-    );
-
-    json!({
-        "status": status,
-        "kind": kind,
-        "display": display,
-        "blocked": blocked,
-        "terminal": terminal,
-        "needs_attention": needs_attention,
-        "next_action": next_action,
-        "approval_primary_action": approval_primary_action,
-        "recovered": recovered,
-        "tool_narrowing_active": tool_narrowing_active,
-        "signals": signals,
-    })
-}
-
-fn unknown_task_status_payload() -> Value {
-    json!({
-        "status": "unknown",
-        "kind": "unknown",
-        "display": "unknown",
-        "blocked": false,
-        "terminal": false,
-        "needs_attention": false,
-        "next_action": "status",
-        "approval_primary_action": Value::Null,
-        "recovered": false,
-        "tool_narrowing_active": false,
-        "signals": [],
-    })
-}
-
-fn derive_task_status_kind(
-    session_state: &str,
-    phase: &str,
-    staleness_state: Option<&str>,
-    cancellation_state: Option<&str>,
-    has_approval_attention: bool,
-    task_progress_status: Option<&str>,
-    task_progress_handle_state: Option<&str>,
-    terminal_outcome_state: Option<&str>,
-    recovery_kind: Option<&str>,
-) -> &'static str {
-    if let Some(task_progress_status) = task_progress_status {
-        return match task_progress_status {
-            "completed" => "completed",
-            "failed" => "failed",
-            "blocked" => {
-                if terminal_outcome_state == Some("present") || recovery_kind.is_some() {
-                    "failed"
-                } else {
-                    "blocked"
-                }
-            }
-            "waiting" => "waiting",
-            "verifying" | "active" => {
-                if task_progress_handle_state == Some("queued") {
-                    "queued"
-                } else if terminal_outcome_state == Some("present") {
-                    match session_state {
-                        "completed" => "completed",
-                        "failed" | "timed_out" => "failed",
-                        _ => "running",
-                    }
-                } else {
-                    "running"
-                }
-            }
-            _ => "running",
-        };
-    }
-
-    if session_state == "completed" {
-        return "completed";
-    }
-
-    if session_state == "failed" {
-        return "failed";
-    }
-
-    if session_state == "timed_out" {
-        return "timed_out";
-    }
-
-    let is_overdue = staleness_state == Some("overdue");
-    if is_overdue {
-        return "overdue";
-    }
-
-    let cancel_requested = cancellation_state == Some("requested");
-    if cancel_requested {
-        return "cancel_requested";
-    }
-
-    if has_approval_attention {
-        return "approval_pending";
-    }
-
-    if session_state == "running" {
-        return "running";
-    }
-
-    let queued_state = session_state == "ready";
-    let queued_phase = phase == "queued";
-    if queued_state || queued_phase {
-        return "queued";
-    }
-
-    "unknown"
-}
-
-fn render_task_status_display(kind: &str, recovered: bool) -> String {
-    let base = kind.to_owned();
-    if !recovered {
-        return base;
-    }
-
-    let display = format!("{base} (recovered)");
-    display
-}
-
-fn task_status_is_blocked(kind: &str) -> bool {
-    matches!(kind, "approval_pending" | "overdue")
-}
-
-fn task_status_is_terminal(kind: &str) -> bool {
-    matches!(kind, "completed" | "failed" | "timed_out")
-}
-
-fn task_status_needs_attention(kind: &str, approval_primary_action: Option<&str>) -> bool {
-    let status_requires_attention = matches!(
-        kind,
-        "approval_pending" | "overdue" | "failed" | "timed_out"
-    );
-    if status_requires_attention {
-        return true;
-    }
-
-    approval_primary_action.is_some()
-}
-
-fn task_status_next_action(kind: &str, approval_primary_action: Option<&str>) -> String {
-    if let Some(approval_primary_action) = approval_primary_action {
-        let next_action = approval_primary_action.to_owned();
-        return next_action;
-    }
-
-    match kind {
-        "approval_pending" => "status".to_owned(),
-        "overdue" => "recover".to_owned(),
-        "queued" => "wait".to_owned(),
-        "running" => "wait".to_owned(),
-        "cancel_requested" => "wait".to_owned(),
-        "completed" => "events".to_owned(),
-        "failed" => "events".to_owned(),
-        "timed_out" => "events".to_owned(),
-        _ => "status".to_owned(),
-    }
-}
-
-fn build_task_status_signals(
-    kind: &str,
-    recovered: bool,
-    tool_narrowing_active: bool,
-    has_approval_attention: bool,
-    staleness_state: Option<&str>,
-    cancellation_state: Option<&str>,
-) -> Vec<String> {
-    let mut signals = Vec::new();
-
-    if has_approval_attention {
-        signals.push("approval_pending".to_owned());
-    }
-
-    if staleness_state == Some("overdue") {
-        signals.push("overdue".to_owned());
-    }
-
-    if cancellation_state == Some("requested") {
-        signals.push("cancel_requested".to_owned());
-    }
-
-    if recovered {
-        signals.push("recovered".to_owned());
-    }
-
-    if tool_narrowing_active {
-        signals.push("tool_narrowing_active".to_owned());
-    }
-
-    let terminal = task_status_is_terminal(kind);
-    if terminal {
-        signals.push("terminal".to_owned());
-    }
-
-    signals
-}
-
-fn recent_events_contains_kind(recent_events: &Value, expected_kind: &str) -> bool {
-    let Some(events) = recent_events.as_array() else {
-        return false;
-    };
-
-    for event in events {
-        let event_kind = event
-            .get("event_kind")
-            .and_then(Value::as_str)
-            .unwrap_or("");
-        let matches_kind = event_kind == expected_kind;
-        if matches_kind {
-            return true;
-        }
-    }
-
-    false
-}
-
-fn primary_approval_action(approval_requests: &Value) -> Option<&str> {
-    let requests = approval_requests.as_array()?;
-
-    for request in requests {
-        let action = request
-            .get("attention")
-            .and_then(|value| value.get("primary_action"))
-            .and_then(Value::as_str);
-        if action.is_some() {
-            return action;
-        }
-    }
-
-    None
-}
-
-fn task_tool_narrowing_active(tool_policy: &Value) -> bool {
-    let effective_tool_ids = tool_policy
-        .get("effective_tool_ids")
-        .and_then(Value::as_array)
-        .cloned()
-        .unwrap_or_default();
-    let base_tool_ids = tool_policy
-        .get("base_tool_ids")
-        .and_then(Value::as_array)
-        .cloned()
-        .unwrap_or_default();
-    let runtime_narrowing = tool_policy.get("effective_runtime_narrowing");
-    let runtime_narrowing = runtime_narrowing.cloned().unwrap_or(Value::Null);
-    let tool_ids_changed = effective_tool_ids != base_tool_ids;
-    let runtime_narrowing_active = !runtime_narrowing.is_null();
-
-    tool_ids_changed || runtime_narrowing_active
 }
 
 fn load_task_status_payload(
@@ -1792,664 +1443,18 @@ fn build_task_next_steps() -> Vec<String> {
 }
 
 fn required_string_field(value: &Value, field: &str, context: &str) -> CliResult<String> {
-    let text = value
+    value
         .get(field)
         .and_then(Value::as_str)
-        .ok_or_else(|| format!("{context} missing string field `{field}`"))?;
-    Ok(text.to_owned())
-}
-
-pub fn render_tasks_cli_text(execution: &TasksCommandExecution) -> CliResult<String> {
-    let command = execution
-        .payload
-        .get("command")
-        .and_then(Value::as_str)
-        .ok_or_else(|| "tasks CLI payload missing command".to_owned())?;
-
-    let rendered = match command {
-        "create" => render_tasks_create_text(&execution.payload)?,
-        "list" => render_tasks_list_text(&execution.payload)?,
-        "status" => render_tasks_status_text(&execution.payload)?,
-        "events" => render_tasks_events_text(&execution.payload)?,
-        "wait" => render_tasks_wait_text(&execution.payload)?,
-        "cancel" | "recover" => render_tasks_mutation_text(&execution.payload)?,
-        other => {
-            return Err(format!("unknown tasks CLI render command `{other}`"));
-        }
-    };
-    Ok(rendered)
-}
-
-fn render_tasks_create_text(payload: &Value) -> CliResult<String> {
-    let task = payload
-        .get("task")
-        .ok_or_else(|| "tasks create payload missing task".to_owned())?;
-    let recipes = payload
-        .get("recipes")
-        .and_then(Value::as_array)
-        .ok_or_else(|| "tasks create payload missing recipes".to_owned())?;
-    let next_steps = payload
-        .get("next_steps")
-        .and_then(Value::as_array)
-        .ok_or_else(|| "tasks create payload missing next_steps".to_owned())?;
-    let scope = payload
-        .get("current_session_id")
-        .and_then(Value::as_str)
-        .unwrap_or("unknown");
-
-    let mut lines = Vec::new();
-    let sanitized_scope = crate::sessions_cli::sanitize_terminal_text(scope);
-    lines.push(format!(
-        "background task queued from scope session `{sanitized_scope}`"
-    ));
-    lines.extend(render_task_detail_lines(task)?);
-    append_task_lookup_error_line(payload, &mut lines);
-
-    if !recipes.is_empty() {
-        for recipe in recipes {
-            let text = recipe.as_str().unwrap_or("");
-            let sanitized_text = crate::sessions_cli::sanitize_terminal_text(text);
-            lines.push(format!("- {sanitized_text}"));
-        }
-    }
-
-    let mut next_lines = Vec::new();
-    if !next_steps.is_empty() {
-        for step in next_steps {
-            let text = step.as_str().unwrap_or("");
-            let sanitized_text = crate::sessions_cli::sanitize_terminal_text(text);
-            next_lines.push(format!("- {sanitized_text}"));
-        }
-    }
-
-    let mut sections = Vec::new();
-    if !next_lines.is_empty() {
-        sections.push(("next steps", next_lines));
-    }
-    sections.push(("queued task", lines));
-    Ok(render_tasks_surface(
-        "task queued",
-        "background tasks",
-        Vec::new(),
-        sections,
-        vec![
-            "Use the next-step commands to inspect, wait on, or cancel the queued task.".to_owned(),
-        ],
-    ))
-}
-
-fn render_tasks_list_text(payload: &Value) -> CliResult<String> {
-    let tasks = payload
-        .get("tasks")
-        .and_then(Value::as_array)
-        .ok_or_else(|| "tasks list payload missing tasks array".to_owned())?;
-    let matched_count = payload
-        .get("matched_count")
-        .and_then(Value::as_u64)
-        .unwrap_or(0);
-    let returned_count = payload
-        .get("returned_count")
-        .and_then(Value::as_u64)
-        .unwrap_or(0);
-    let scope = payload
-        .get("current_session_id")
-        .and_then(Value::as_str)
-        .unwrap_or("unknown");
-
-    let mut lines = Vec::new();
-    let sanitized_scope = crate::sessions_cli::sanitize_terminal_text(scope);
-    lines.push(format!(
-        "visible background tasks from scope session `{sanitized_scope}`: {returned_count}/{matched_count}"
-    ));
-    if tasks.is_empty() {
-        lines.push("No async background tasks are currently visible.".to_owned());
-        return Ok(render_tasks_surface(
-            "visible tasks",
-            "background tasks",
-            Vec::new(),
-            vec![("tasks", lines)],
-            vec![
-                "Use `tasks create` to queue a new background delegate from the current session."
-                    .to_owned(),
-            ],
-        ));
-    }
-
-    for task in tasks {
-        let line = render_task_brief_line(task)?;
-        lines.push(format!("- {line}"));
-    }
-
-    Ok(render_tasks_surface(
-        "visible tasks",
-        "background tasks",
-        Vec::new(),
-        vec![("tasks", lines)],
-        vec![
-            "Use `tasks status <id>` for one task or `tasks wait <id>` to follow it incrementally."
-                .to_owned(),
-        ],
-    ))
-}
-
-fn render_tasks_status_text(payload: &Value) -> CliResult<String> {
-    let task = payload
-        .get("task")
-        .ok_or_else(|| "tasks status payload missing task".to_owned())?;
-    let lines = render_task_detail_lines(task)?;
-    Ok(render_tasks_surface(
-        "task detail",
-        "background tasks",
-        Vec::new(),
-        vec![("task", lines)],
-        vec![
-            "Use `tasks events <id>` or `tasks wait <id>` to keep inspecting the task lifecycle."
-                .to_owned(),
-        ],
-    ))
-}
-
-fn render_tasks_events_text(payload: &Value) -> CliResult<String> {
-    let task_id = payload
-        .get("task_id")
-        .and_then(Value::as_str)
-        .unwrap_or("unknown");
-    let events = payload
-        .get("events")
-        .and_then(Value::as_array)
-        .ok_or_else(|| "tasks events payload missing events array".to_owned())?;
-    let next_after_id = payload
-        .get("next_after_id")
-        .and_then(Value::as_i64)
-        .unwrap_or(0);
-
-    let mut lines = Vec::new();
-    let sanitized_task_id = crate::sessions_cli::sanitize_terminal_text(task_id);
-    lines.push(format!(
-        "events for `{sanitized_task_id}` (next_after_id={next_after_id})"
-    ));
-    if events.is_empty() {
-        lines.push("No newer events.".to_owned());
-        return Ok(render_tasks_surface(
-            "task events",
-            "background tasks",
-            Vec::new(),
-            vec![("events", lines)],
-            vec!["Use `tasks wait <id>` to continue following this task.".to_owned()],
-        ));
-    }
-
-    for event in events {
-        let event_id = event.get("id").and_then(Value::as_i64).unwrap_or_default();
-        let event_kind = event
-            .get("event_kind")
-            .and_then(Value::as_str)
-            .unwrap_or("unknown");
-        let ts = event.get("ts").and_then(Value::as_i64).unwrap_or_default();
-        let sanitized_event_kind = crate::sessions_cli::sanitize_terminal_text(event_kind);
-        lines.push(format!("- #{event_id} {sanitized_event_kind} ts={ts}"));
-    }
-
-    Ok(render_tasks_surface(
-        "task events",
-        "background tasks",
-        Vec::new(),
-        vec![("events", lines)],
-        vec!["Use `tasks wait <id>` to continue following this task.".to_owned()],
-    ))
-}
-
-fn render_tasks_wait_text(payload: &Value) -> CliResult<String> {
-    let wait_status = payload
-        .get("wait_status")
-        .and_then(Value::as_str)
-        .unwrap_or("unknown");
-    let task = payload
-        .get("task")
-        .ok_or_else(|| "tasks wait payload missing task".to_owned())?;
-    let events = payload
-        .get("events")
-        .and_then(Value::as_array)
-        .ok_or_else(|| "tasks wait payload missing events array".to_owned())?;
-    let next_after_id = payload
-        .get("next_after_id")
-        .and_then(Value::as_i64)
-        .unwrap_or(0);
-
-    let mut lines = Vec::new();
-    lines.push(format!(
-        "wait result: {wait_status} (next_after_id={next_after_id})"
-    ));
-    lines.extend(render_task_detail_lines(task)?);
-    if !events.is_empty() {
-        lines.push("observed events:".to_owned());
-        for event in events {
-            let event_id = event.get("id").and_then(Value::as_i64).unwrap_or_default();
-            let event_kind = event
-                .get("event_kind")
-                .and_then(Value::as_str)
-                .unwrap_or("unknown");
-            let sanitized_event_kind = crate::sessions_cli::sanitize_terminal_text(event_kind);
-            lines.push(format!("- #{event_id} {sanitized_event_kind}"));
-        }
-    }
-
-    Ok(render_tasks_surface(
-        "task wait",
-        "background tasks",
-        Vec::new(),
-        vec![("result", lines)],
-        vec!["Re-run `tasks wait` with the returned cursor when you need more updates.".to_owned()],
-    ))
-}
-
-fn render_tasks_mutation_text(payload: &Value) -> CliResult<String> {
-    let command = payload
-        .get("command")
-        .and_then(Value::as_str)
-        .unwrap_or("unknown");
-    let task = payload
-        .get("task")
-        .ok_or_else(|| "tasks mutation payload missing task".to_owned())?;
-    let action = payload.get("action").cloned().unwrap_or(Value::Null);
-    let dry_run = payload
-        .get("dry_run")
-        .and_then(Value::as_bool)
-        .unwrap_or(false);
-    let result = payload.get("result").and_then(Value::as_str);
-    let message = payload.get("message").and_then(Value::as_str);
-
-    let mut lines = Vec::new();
-    lines.push(format!("{command} dry_run={dry_run}"));
-    if let Some(result) = result {
-        let sanitized_result = crate::sessions_cli::sanitize_terminal_text(result);
-        lines.push(format!("result: {sanitized_result}"));
-    }
-    if let Some(message) = message {
-        let sanitized_message = crate::sessions_cli::sanitize_terminal_text(message);
-        lines.push(format!("message: {sanitized_message}"));
-    }
-    if !action.is_null() {
-        let rendered_action = serde_json::to_string_pretty(&action)
-            .map_err(|error| format!("render action failed: {error}"))?;
-        lines.push("action:".to_owned());
-        lines.push(rendered_action);
-    }
-    lines.extend(render_task_detail_lines(task)?);
-    append_task_lookup_error_line(payload, &mut lines);
-    Ok(render_tasks_surface(
-        "task action",
-        "background tasks",
-        Vec::new(),
-        vec![("action result", lines)],
-        vec!["Use `tasks status <id>` to verify the task state after the action.".to_owned()],
-    ))
-}
-
-fn render_tasks_surface(
-    title: &str,
-    subtitle: &str,
-    intro_lines: Vec<String>,
-    sections: Vec<(&str, Vec<String>)>,
-    footer_lines: Vec<String>,
-) -> String {
-    let sections = sections
-        .into_iter()
-        .map(
-            |(section_title, lines)| mvp::tui_surface::TuiSectionSpec::Narrative {
-                title: Some(section_title.to_owned()),
-                lines,
-            },
-        )
-        .collect();
-    let screen = mvp::tui_surface::TuiScreenSpec {
-        header_style: mvp::tui_surface::TuiHeaderStyle::Compact,
-        subtitle: Some(subtitle.to_owned()),
-        title: Some(title.to_owned()),
-        progress_line: None,
-        intro_lines,
-        sections,
-        choices: Vec::new(),
-        footer_lines,
-    };
-    mvp::tui_surface::render_tui_screen_spec_ratatui(
-        &screen,
-        mvp::presentation::detect_render_width(),
-        false,
-    )
-    .join("\n")
-}
-
-fn render_task_brief_line(task: &Value) -> CliResult<String> {
-    let task_id = required_string_field(task, "task_id", "task summary")?;
-    let state = task
-        .get("session_state")
-        .and_then(Value::as_str)
-        .unwrap_or("unknown");
-    let phase = task
-        .get("phase")
-        .and_then(Value::as_str)
-        .unwrap_or("unknown");
-    let task_status = task
-        .get("task_status")
-        .cloned()
-        .unwrap_or_else(unknown_task_status_payload);
-    let status_display = task_status
-        .get("display")
-        .and_then(Value::as_str)
-        .unwrap_or("unknown");
-    let blocked = task_status
-        .get("blocked")
-        .and_then(Value::as_bool)
-        .unwrap_or(false);
-    let workflow_phase = task
-        .get("workflow")
-        .and_then(|value| value.get("phase"))
-        .and_then(Value::as_str)
-        .unwrap_or("unknown");
-    let label = task.get("label").and_then(Value::as_str).unwrap_or("-");
-    let approval_attention = task
-        .get("approval")
-        .and_then(|value| value.get("attention_summary"))
-        .and_then(|value| value.get("needs_attention_count"))
-        .and_then(Value::as_u64)
-        .unwrap_or(0);
-    let owner_kind = task
-        .get("owner_kind")
-        .and_then(Value::as_str)
-        .unwrap_or("unknown");
-    let signals = task_status
-        .get("signals")
-        .and_then(Value::as_array)
-        .map(|values| render_string_array(values))
-        .unwrap_or_else(|| "-".to_owned());
-    let sanitized_task_id = crate::sessions_cli::sanitize_terminal_text(task_id.as_str());
-    let sanitized_label = crate::sessions_cli::sanitize_terminal_text(label);
-    let sanitized_owner_kind = crate::sessions_cli::sanitize_terminal_text(owner_kind);
-    let line = format!(
-        "{sanitized_task_id} status={status_display} blocked={blocked} state={state} workflow_phase={workflow_phase} delegate_phase={phase} label={sanitized_label} owner_kind={sanitized_owner_kind} approval_attention={approval_attention} signals={signals}"
-    );
-    Ok(line)
-}
-
-fn render_task_detail_lines(task: &Value) -> CliResult<Vec<String>> {
-    let task_id = required_string_field(task, "task_id", "task detail")?;
-    let task_session_id = task
-        .get("task_session_id")
-        .and_then(Value::as_str)
-        .unwrap_or("-");
-    let owner_session_id = task
-        .get("owner_session_id")
-        .and_then(Value::as_str)
-        .unwrap_or("-");
-    let scope_session_id = task
-        .get("scope_session_id")
-        .and_then(Value::as_str)
-        .unwrap_or("unknown");
-    let task_status = task
-        .get("task_status")
-        .cloned()
-        .unwrap_or_else(unknown_task_status_payload);
-    let task_status_display = task_status
-        .get("display")
-        .and_then(Value::as_str)
-        .unwrap_or("unknown");
-    let blocked = task_status
-        .get("blocked")
-        .and_then(Value::as_bool)
-        .unwrap_or(false);
-    let needs_attention = task_status
-        .get("needs_attention")
-        .and_then(Value::as_bool)
-        .unwrap_or(false);
-    let next_action = task_status
-        .get("next_action")
-        .and_then(Value::as_str)
-        .unwrap_or("status");
-    let task_signals = task_status
-        .get("signals")
-        .and_then(Value::as_array)
-        .map(|values| render_string_array(values))
-        .unwrap_or_else(|| "-".to_owned());
-    let label = task.get("label").and_then(Value::as_str).unwrap_or("-");
-    let state = task
-        .get("session_state")
-        .and_then(Value::as_str)
-        .unwrap_or("unknown");
-    let phase = task
-        .get("phase")
-        .and_then(Value::as_str)
-        .unwrap_or("unknown");
-    let owner_kind = task
-        .get("owner_kind")
-        .and_then(Value::as_str)
-        .unwrap_or("unknown");
-    let workflow_id = task
-        .get("workflow")
-        .and_then(|value| value.get("workflow_id"))
-        .and_then(Value::as_str)
-        .unwrap_or("-");
-    let workflow_phase = task
-        .get("workflow")
-        .and_then(|value| value.get("phase"))
-        .and_then(Value::as_str)
-        .unwrap_or("-");
-    let workflow_operation_kind = task
-        .get("workflow")
-        .and_then(|value| value.get("operation_kind"))
-        .and_then(Value::as_str)
-        .unwrap_or("-");
-    let workflow_operation_scope = task
-        .get("workflow")
-        .and_then(|value| value.get("operation_scope"))
-        .and_then(Value::as_str)
-        .unwrap_or("-");
-    let workflow_task_session_id = task
-        .get("workflow")
-        .and_then(|value| value.get("task_session_id"))
-        .and_then(Value::as_str)
-        .unwrap_or("-");
-    let workflow_binding_mode = task
-        .get("workflow")
-        .and_then(|value| value.get("binding"))
-        .and_then(|value| value.get("mode"))
-        .and_then(Value::as_str)
-        .unwrap_or("-");
-    let workflow_execution_surface = task
-        .get("workflow")
-        .and_then(|value| value.get("binding"))
-        .and_then(|value| value.get("execution_surface"))
-        .and_then(Value::as_str)
-        .unwrap_or("-");
-    let workflow_worktree_id = task
-        .get("workflow")
-        .and_then(|value| value.get("binding"))
-        .and_then(|value| value.get("worktree"))
-        .and_then(|value| value.get("worktree_id"))
-        .and_then(Value::as_str)
-        .unwrap_or("-");
-    let workflow_workspace_root = task
-        .get("workflow")
-        .and_then(|value| value.get("binding"))
-        .and_then(|value| value.get("worktree"))
-        .and_then(|value| value.get("workspace_root"))
-        .and_then(Value::as_str)
-        .unwrap_or("-");
-    let timeout_seconds = task
-        .get("timeout_seconds")
-        .and_then(Value::as_u64)
-        .map(|value| value.to_string())
-        .unwrap_or_else(|| "unknown".to_owned());
-    let last_error = task
-        .get("last_error")
-        .and_then(Value::as_str)
-        .unwrap_or("-");
-    let approval_total = task
-        .get("approval")
-        .and_then(|value| value.get("matched_count"))
-        .and_then(Value::as_u64)
-        .unwrap_or(0);
-    let approval_attention = task
-        .get("approval")
-        .and_then(|value| value.get("attention_summary"))
-        .and_then(|value| value.get("needs_attention_count"))
-        .and_then(Value::as_u64)
-        .unwrap_or(0);
-    let requested_tool_ids =
-        render_task_tool_policy_tool_ids(task, "visible_requested_tool_ids", "requested_tool_ids");
-    let effective_tool_ids =
-        render_task_tool_policy_tool_ids(task, "visible_effective_tool_ids", "effective_tool_ids");
-    let effective_runtime_narrowing = task
-        .get("tool_policy")
-        .and_then(|value| value.get("effective_runtime_narrowing"))
-        .cloned()
-        .unwrap_or(Value::Null);
-    let prompt_frame_summary =
-        crate::session_prompt_frame_cli::render_prompt_frame_summary(task.get("prompt_frame"));
-    let safe_lane_summary =
-        crate::session_runtime_truth_cli::render_safe_lane_summary(task.get("safe_lane"));
-    let turn_checkpoint_summary = crate::session_runtime_truth_cli::render_turn_checkpoint_summary(
-        task.get("turn_checkpoint"),
-    );
-    let rendered_runtime_narrowing = if effective_runtime_narrowing.is_null() {
-        "-".to_owned()
-    } else {
-        serde_json::to_string(&effective_runtime_narrowing)
-            .map_err(|error| format!("render runtime narrowing failed: {error}"))?
-    };
-    let sanitized_task_id = crate::sessions_cli::sanitize_terminal_text(task_id.as_str());
-    let sanitized_task_session_id = crate::sessions_cli::sanitize_terminal_text(task_session_id);
-    let sanitized_owner_session_id = crate::sessions_cli::sanitize_terminal_text(owner_session_id);
-    let sanitized_scope_session_id = crate::sessions_cli::sanitize_terminal_text(scope_session_id);
-    let sanitized_label = crate::sessions_cli::sanitize_terminal_text(label);
-    let sanitized_last_error = crate::sessions_cli::sanitize_terminal_text(last_error);
-    let sanitized_requested_tool_ids =
-        crate::sessions_cli::sanitize_terminal_text(requested_tool_ids.as_str());
-    let sanitized_effective_tool_ids =
-        crate::sessions_cli::sanitize_terminal_text(effective_tool_ids.as_str());
-    let sanitized_runtime_narrowing =
-        crate::sessions_cli::sanitize_terminal_text(rendered_runtime_narrowing.as_str());
-    let sanitized_prompt_frame_summary =
-        crate::sessions_cli::sanitize_terminal_text(prompt_frame_summary.as_str());
-    let sanitized_safe_lane_summary =
-        crate::sessions_cli::sanitize_terminal_text(safe_lane_summary.as_str());
-    let sanitized_turn_checkpoint_summary =
-        crate::sessions_cli::sanitize_terminal_text(turn_checkpoint_summary.as_str());
-    let approval_lookup_error = task
-        .get("approval_lookup_error")
-        .and_then(Value::as_str)
-        .map(crate::sessions_cli::sanitize_terminal_text);
-    let tool_policy_lookup_error = task
-        .get("tool_policy_lookup_error")
-        .and_then(Value::as_str)
-        .map(crate::sessions_cli::sanitize_terminal_text);
-
-    let mut lines = Vec::new();
-    lines.push(format!("task_id: {sanitized_task_id}"));
-    lines.push(format!("task_session_id: {sanitized_task_session_id}"));
-    lines.push(format!("owner_session_id: {sanitized_owner_session_id}"));
-    lines.push(format!("scope_session_id: {sanitized_scope_session_id}"));
-    lines.push(format!("label: {sanitized_label}"));
-    lines.push(format!("task_status: {task_status_display}"));
-    lines.push(format!("task_blocked: {blocked}"));
-    lines.push(format!("task_needs_attention: {needs_attention}"));
-    lines.push(format!("task_next_action: {next_action}"));
-    lines.push(format!("task_signals: {task_signals}"));
-    lines.push(format!("state: {state}"));
-    lines.push(format!("workflow_id: {workflow_id}"));
-    lines.push(format!("workflow_phase: {workflow_phase}"));
-    lines.push(format!(
-        "workflow_operation_kind: {workflow_operation_kind}"
-    ));
-    lines.push(format!(
-        "workflow_operation_scope: {workflow_operation_scope}"
-    ));
-    lines.push(format!(
-        "workflow_task_session_id: {workflow_task_session_id}"
-    ));
-    lines.push(format!("workflow_binding_mode: {workflow_binding_mode}"));
-    lines.push(format!(
-        "workflow_execution_surface: {workflow_execution_surface}"
-    ));
-    lines.push(format!("workflow_worktree_id: {workflow_worktree_id}"));
-    lines.push(format!(
-        "workflow_workspace_root: {workflow_workspace_root}"
-    ));
-    lines.push(format!("phase: {phase}"));
-    lines.push(format!("owner_kind: {owner_kind}"));
-    lines.push(format!("timeout_seconds: {timeout_seconds}"));
-    lines.push(format!("last_error: {sanitized_last_error}"));
-    lines.push(format!("approval_requests: {approval_total}"));
-    lines.push(format!("approval_attention: {approval_attention}"));
-    if requested_tool_ids != "-" {
-        lines.push(format!(
-            "requested_tool_ids: {sanitized_requested_tool_ids}"
-        ));
-    }
-    lines.push(format!(
-        "effective_tool_ids: {sanitized_effective_tool_ids}"
-    ));
-    lines.push(format!(
-        "effective_runtime_narrowing: {sanitized_runtime_narrowing}"
-    ));
-    lines.push(format!("prompt_frame: {sanitized_prompt_frame_summary}"));
-    lines.push(format!("safe_lane: {sanitized_safe_lane_summary}"));
-    lines.push(format!(
-        "turn_checkpoint: {sanitized_turn_checkpoint_summary}"
-    ));
-    if let Some(approval_lookup_error) = approval_lookup_error {
-        lines.push(format!("approval_lookup_error: {approval_lookup_error}"));
-    }
-    if let Some(tool_policy_lookup_error) = tool_policy_lookup_error {
-        lines.push(format!(
-            "tool_policy_lookup_error: {tool_policy_lookup_error}"
-        ));
-    }
-    Ok(lines)
-}
-
-fn append_task_lookup_error_line(payload: &Value, lines: &mut Vec<String>) {
-    let Some(task_lookup_error) = payload.get("task_lookup_error").and_then(Value::as_str) else {
-        return;
-    };
-    let sanitized_task_lookup_error =
-        crate::sessions_cli::sanitize_terminal_text(task_lookup_error);
-    lines.push(format!("task_lookup_error: {sanitized_task_lookup_error}"));
-}
-
-fn render_task_tool_policy_tool_ids(task: &Value, visible_field: &str, raw_field: &str) -> String {
-    task.get("tool_policy")
-        .and_then(|value| value.get(visible_field))
-        .and_then(Value::as_array)
-        .filter(|values| !values.is_empty())
-        .map(|values| render_string_array(values))
-        .or_else(|| {
-            task.get("tool_policy")
-                .and_then(|value| value.get(raw_field))
-                .and_then(Value::as_array)
-                .filter(|values| !values.is_empty())
-                .map(|values| render_string_array(values))
-        })
-        .unwrap_or_else(|| "-".to_owned())
-}
-
-fn render_string_array(values: &[Value]) -> String {
-    let mut items = Vec::new();
-    for value in values {
-        if let Some(text) = value.as_str() {
-            items.push(text.to_owned());
-        }
-    }
-    if items.is_empty() {
-        return "-".to_owned();
-    }
-    items.join(", ")
+        .map(str::to_owned)
+        .ok_or_else(|| format!("{context} missing string field `{field}`"))
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::mvp;
+    use crate::tasks_cli::render_support::{render_task_brief_line, render_task_detail_lines};
     use mvp::session::repository::{
         NewSessionEvent, NewSessionRecord, SessionKind, SessionRepository, SessionState,
     };
